pyproject.toml
[tool.poetry]
name = "sql-alchemy"
version = "0.1.0"
description = ""
authors = ["Evgeniy Polyakov <evgenickus@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
sqlalchemy = "^2.0.30"
fastapi = "^0.111.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
